+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Named handles for the GBuffer textures
+
+### Changed
+
+- GBuffer shutdown method now uses texture manager directly
+
 ## [0.0.4] - 2025-05-07
 
 ### Added